Created actuators function
